requirements.txt
aiomysql==0.1.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
Authlib==1.2.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
decorator==4.4.2
dnspython==2.4.2
dynaconf==3.2.4
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.105.0
fastapi-sa==0.1.0
greenlet==3.0.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
moviepy==1.0.3
numpy==1.26.2
orjson==3.9.10
passlib==1.7.4
Pillow==10.1.0
proglog==0.1.10
psycopg2==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.21.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
